# dev/iot-rest-api/api/models/Sensor.js
class Sensor {
  constructor(
    id,
    documentationLink,
    dateCreated,
    dateUpdated,
    active,
    refreshInterval,
  ) {
    this.id = id;
    this.documentationLink = documentationLink;
    this.dateCreated = dateCreated;
    this.dateUpdated = dateUpdated;
    this.active = active;
    this.refreshInterval = refreshInterval;
  }
}

module.exports = Sensor;